+++ common/cpp/src/flutter_frame_cryptor.cc
@@ -4,14 +4,14 @@
 
 namespace flutter_webrtc_plugin {
 
-libwebrtc::Algorithm AlgorithmFromInt(int algorithm) {
+libwebrtc::FrameCryptorAlgorithm AlgorithmFromInt(int algorithm) {
   switch (algorithm) {
     case 0:
-      return libwebrtc::Algorithm::kAesGcm;
+      return libwebrtc::FrameCryptorAlgorithm::kAesGcm;
     case 1:
-      return libwebrtc::Algorithm::kAesCbc;
+      return libwebrtc::FrameCryptorAlgorithm::kAesCbc;
     default:
-      return libwebrtc::Algorithm::kAesGcm;
+      return libwebrtc::FrameCryptorAlgorithm::kAesGcm;
   }
 }
 
@@ -40,7 +40,8 @@ void FlutterFrameCryptorObserver::OnFrameCryptionStateChanged(
     libwebrtc::RTCFrameCryptionState state) {
   EncodableMap params;
   params[EncodableValue("event")] = EncodableValue("frameCryptionStateChanged");
-  params[EncodableValue("participantId")] = EncodableValue(participant_id.std_string());
+  params[EncodableValue("participantId")] =
+      EncodableValue(participant_id.std_string());
   params[EncodableValue("state")] =
       EncodableValue(frameCryptionStateToString(state));
   event_channel_->Success(EncodableValue(params));
@@ -49,7 +50,7 @@ void FlutterFrameCryptorObserver::OnFrameCryptionStateChanged(
 bool FlutterFrameCryptor::HandleFrameCryptorMethodCall(
     const MethodCallProxy& method_call,
     std::unique_ptr<MethodResultProxy> result,
-    std::unique_ptr<MethodResultProxy> *outResult) {
+    std::unique_ptr<MethodResultProxy>* outResult) {
   const std::string& method_name = method_call.method_name();
   if (!method_call.arguments()) {
     result->Error("Bad Arguments", "Null arguments received");
@@ -87,7 +88,7 @@ bool FlutterFrameCryptor::HandleFrameCryptorMethodCall(
   } else if (method_name == "keyProviderExportSharedKey") {
     KeyProviderExportSharedKey(params, std::move(result));
     return true;
-  }else if (method_name == "keyProviderSetKey") {
+  } else if (method_name == "keyProviderSetKey") {
     KeyProviderSetKey(params, std::move(result));
     return true;
   } else if (method_name == "keyProviderRatchetKey") {
@@ -96,14 +97,14 @@ bool FlutterFrameCryptor::HandleFrameCryptorMethodCall(
   } else if (method_name == "keyProviderExportKey") {
     KeyProviderExportKey(params, std::move(result));
     return true;
-  }  else if (method_name == "keyProviderSetSifTrailer") {
+  } else if (method_name == "keyProviderSetSifTrailer") {
     KeyProviderSetSifTrailer(params, std::move(result));
     return true;
   } else if (method_name == "keyProviderDispose") {
     KeyProviderDispose(params, std::move(result));
     return true;
   }
-  
+
   *outResult = std::move(result);
   return false;
 }
@@ -154,19 +155,21 @@ void FlutterFrameCryptor::FrameCryptorFactoryCreateFrameCryptor(
       return;
     }
     std::string uuid = base_->GenerateUUID();
-    auto keyProvider = key_providers_[keyProviderId];
+    auto keyProvider = base_->key_providers_[keyProviderId];
     if (keyProvider == nullptr) {
       result->Error("FrameCryptorFactoryCreateFrameCryptorFailed",
                     "keyProvider is null");
       return;
     }
     auto frameCryptor =
-        libwebrtc::FrameCryptorFactory::frameCryptorFromRtpSender(base_->factory_,
-            string(participantId), sender, AlgorithmFromInt(algorithm),
-            keyProvider);
+        libwebrtc::FrameCryptorFactory::frameCryptorFromRtpSender(
+            base_->factory_, string(participantId), sender,
+            AlgorithmFromInt(algorithm), keyProvider);
     std::string event_channel = "FlutterWebRTC/frameCryptorEvent" + uuid;
 
-    scoped_refptr<FlutterFrameCryptorObserver> observer(new RefCountedObject<FlutterFrameCryptorObserver>(base_->messenger_, base_->task_runner_, event_channel));
+    scoped_refptr<FlutterFrameCryptorObserver> observer(
+        new RefCountedObject<FlutterFrameCryptorObserver>(
+            base_->messenger_, base_->task_runner_, event_channel));
 
     frameCryptor->RegisterRTCFrameCryptorObserver(observer);
 
@@ -184,15 +187,17 @@ void FlutterFrameCryptor::FrameCryptorFactoryCreateFrameCryptor(
       return;
     }
     std::string uuid = base_->GenerateUUID();
-    auto keyProvider = key_providers_[keyProviderId];
+    auto keyProvider = base_->key_providers_[keyProviderId];
     auto frameCryptor =
-        libwebrtc::FrameCryptorFactory::frameCryptorFromRtpReceiver(base_->factory_,
-            string(participantId), receiver, AlgorithmFromInt(algorithm),
-            keyProvider);
+        libwebrtc::FrameCryptorFactory::frameCryptorFromRtpReceiver(
+            base_->factory_, string(participantId), receiver,
+            AlgorithmFromInt(algorithm), keyProvider);
 
     std::string event_channel = "FlutterWebRTC/frameCryptorEvent" + uuid;
 
-    scoped_refptr<FlutterFrameCryptorObserver> observer(new RefCountedObject<FlutterFrameCryptorObserver>(base_->messenger_, base_->task_runner_, event_channel));
+    scoped_refptr<FlutterFrameCryptorObserver> observer(
+        new RefCountedObject<FlutterFrameCryptorObserver>(
+            base_->messenger_, base_->task_runner_, event_channel));
 
     frameCryptor->RegisterRTCFrameCryptorObserver(observer.get());
 
@@ -309,19 +314,19 @@ void FlutterFrameCryptor::FrameCryptorFactoryCreateKeyProvider(
     const EncodableMap& constraints,
     std::unique_ptr<MethodResultProxy> result) {
   libwebrtc::KeyProviderOptions options;
-  
 
   auto keyProviderOptions = findMap(constraints, "keyProviderOptions");
   if (keyProviderOptions == EncodableMap()) {
-    result->Error("FrameCryptorFactoryCreateKeyProviderFailed", "keyProviderOptions is null");
+    result->Error("FrameCryptorFactoryCreateKeyProviderFailed",
+                  "keyProviderOptions is null");
     return;
   }
 
   auto sharedKey = findBoolean(keyProviderOptions, "sharedKey");
   options.shared_key = sharedKey;
 
-
-  auto uncryptedMagicBytes = findVector(keyProviderOptions, "uncryptedMagicBytes");
+  auto uncryptedMagicBytes =
+      findVector(keyProviderOptions, "uncryptedMagicBytes");
   if (uncryptedMagicBytes.size() != 0) {
     options.uncrypted_magic_bytes = uncryptedMagicBytes;
   }
@@ -336,7 +341,7 @@ void FlutterFrameCryptor::FrameCryptorFactoryCreateKeyProvider(
   options.ratchet_salt = ratchetSalt;
 
   auto ratchetWindowSize = findInt(keyProviderOptions, "ratchetWindowSize");
-  if (ratchetWindowSize  == -1) {
+  if (ratchetWindowSize == -1) {
     result->Error("FrameCryptorFactoryCreateKeyProviderFailed",
                   "ratchetSalt is null");
     return;
@@ -350,8 +355,10 @@ void FlutterFrameCryptor::FrameCryptorFactoryCreateKeyProvider(
   auto keyRingSize = findInt(keyProviderOptions, "keyRingSize");
   options.key_ring_size = keyRingSize;
 
-  auto discardFrameWhenCryptorNotReady = findBoolean(keyProviderOptions, "discardFrameWhenCryptorNotReady");
-  options.discard_frame_when_cryptor_not_ready = discardFrameWhenCryptorNotReady;
+  auto discardFrameWhenCryptorNotReady =
+      findBoolean(keyProviderOptions, "discardFrameWhenCryptorNotReady");
+  options.discard_frame_when_cryptor_not_ready =
+      discardFrameWhenCryptorNotReady;
 
   auto keyProvider = libwebrtc::KeyProvider::Create(&options);
   if (nullptr == keyProvider.get()) {
@@ -360,21 +367,22 @@ void FlutterFrameCryptor::FrameCryptorFactoryCreateKeyProvider(
     return;
   }
   auto uuid = base_->GenerateUUID();
-  key_providers_[uuid] = keyProvider;
+  base_->key_providers_[uuid] = keyProvider;
   EncodableMap params;
   params[EncodableValue("keyProviderId")] = uuid;
   result->Success(EncodableValue(params));
 }
 
-void FlutterFrameCryptor::KeyProviderSetSharedKey(const EncodableMap& constraints,
-                      std::unique_ptr<MethodResultProxy> result) {
+void FlutterFrameCryptor::KeyProviderSetSharedKey(
+    const EncodableMap& constraints,
+    std::unique_ptr<MethodResultProxy> result) {
   auto keyProviderId = findString(constraints, "keyProviderId");
   if (keyProviderId == std::string()) {
     result->Error("KeyProviderSetSharedKeyFailed", "keyProviderId is null");
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderSetSharedKeyFailed", "keyProvider is null");
     return;
@@ -392,22 +400,22 @@ void FlutterFrameCryptor::KeyProviderSetSharedKey(const EncodableMap& constraint
     return;
   }
 
-
   keyProvider->SetSharedKey(key_index, vector<uint8_t>(key));
   EncodableMap params;
   params[EncodableValue("result")] = true;
   result->Success(EncodableValue(params));
 }
 
-void FlutterFrameCryptor::KeyProviderRatchetSharedKey(const EncodableMap& constraints,
-                       std::unique_ptr<MethodResultProxy> result) {
- auto keyProviderId = findString(constraints, "keyProviderId");
+void FlutterFrameCryptor::KeyProviderRatchetSharedKey(
+    const EncodableMap& constraints,
+    std::unique_ptr<MethodResultProxy> result) {
+  auto keyProviderId = findString(constraints, "keyProviderId");
   if (keyProviderId == std::string()) {
     result->Error("KeyProviderRatchetSharedKeyFailed", "keyProviderId is null");
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderRatchetSharedKeyFailed", "keyProvider is null");
     return;
@@ -426,15 +434,16 @@ void FlutterFrameCryptor::KeyProviderRatchetSharedKey(const EncodableMap& constr
   result->Success(EncodableValue(params));
 }
 
-void FlutterFrameCryptor::KeyProviderExportSharedKey(const EncodableMap& constraints,
-                      std::unique_ptr<MethodResultProxy> result) {
-auto keyProviderId = findString(constraints, "keyProviderId");
+void FlutterFrameCryptor::KeyProviderExportSharedKey(
+    const EncodableMap& constraints,
+    std::unique_ptr<MethodResultProxy> result) {
+  auto keyProviderId = findString(constraints, "keyProviderId");
   if (keyProviderId == std::string()) {
     result->Error("KeyProviderExportSharedKeyFailed", "keyProviderId is null");
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderExportSharedKeyFailed", "keyProvider is null");
     return;
@@ -453,15 +462,16 @@ auto keyProviderId = findString(constraints, "keyProviderId");
   result->Success(EncodableValue(params));
 }
 
-void FlutterFrameCryptor::KeyProviderExportKey(const EncodableMap& constraints,
-                      std::unique_ptr<MethodResultProxy> result) {
+void FlutterFrameCryptor::KeyProviderExportKey(
+    const EncodableMap& constraints,
+    std::unique_ptr<MethodResultProxy> result) {
   auto keyProviderId = findString(constraints, "keyProviderId");
   if (keyProviderId == std::string()) {
     result->Error("KeyProviderExportKeyFailed", "keyProviderId is null");
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderExportKeyFailed", "keyProvider is null");
     return;
@@ -486,15 +496,16 @@ void FlutterFrameCryptor::KeyProviderExportKey(const EncodableMap& constraints,
   result->Success(EncodableValue(params));
 }
 
-void FlutterFrameCryptor::KeyProviderSetSifTrailer(const EncodableMap& constraints,
-                       std::unique_ptr<MethodResultProxy> result) {
+void FlutterFrameCryptor::KeyProviderSetSifTrailer(
+    const EncodableMap& constraints,
+    std::unique_ptr<MethodResultProxy> result) {
   auto keyProviderId = findString(constraints, "keyProviderId");
   if (keyProviderId == std::string()) {
     result->Error("KeyProviderSetSifTrailerFailed", "keyProviderId is null");
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderSetSifTrailerFailed", "keyProvider is null");
     return;
@@ -521,7 +532,7 @@ void FlutterFrameCryptor::KeyProviderSetKey(
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderSetKeyFailed", "keyProvider is null");
     return;
@@ -559,7 +570,7 @@ void FlutterFrameCryptor::KeyProviderRatchetKey(
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderSetKeysFailed", "keyProvider is null");
     return;
@@ -584,7 +595,6 @@ void FlutterFrameCryptor::KeyProviderRatchetKey(
   result->Success(EncodableValue(params));
 }
 
-
 void FlutterFrameCryptor::KeyProviderDispose(
     const EncodableMap& constraints,
     std::unique_ptr<MethodResultProxy> result) {
@@ -594,12 +604,12 @@ void FlutterFrameCryptor::KeyProviderDispose(
     return;
   }
 
-  auto keyProvider = key_providers_[keyProviderId];
+  auto keyProvider = base_->key_providers_[keyProviderId];
   if (nullptr == keyProvider.get()) {
     result->Error("KeyProviderDisposeFailed", "keyProvider is null");
     return;
   }
-  key_providers_.erase(keyProviderId);
+  base_->key_providers_.erase(keyProviderId);
   EncodableMap params;
   params[EncodableValue("result")] = "success";
   result->Success(EncodableValue(params));