+++ common/darwin/Classes/FlutterRTCFrameCryptor.m
@@ -88,6 +88,17 @@
   }
 }
 
+- (RTCKeyDerivationAlgorithm)getKeyDerivationAlgorithm:(NSNumber*)algorithm {
+  switch ([algorithm intValue]) {
+    case 0:
+      return RTCKeyDerivationAlgorithmPBKDF2;
+    case 1:
+      return RTCKeyDerivationAlgorithmHKDF;
+    default:
+      return RTCKeyDerivationAlgorithmPBKDF2;
+  }
+}
+
 - (void)frameCryptorFactoryCreateFrameCryptor:(nonnull NSDictionary*)constraints
                                        result:(nonnull FlutterResult)result {
   NSString* peerConnectionId = constraints[@"peerConnectionId"];
@@ -353,6 +364,8 @@
   NSNumber* keyRingSize = keyProviderOptions[@"keyRingSize"];
 
   NSNumber* discardFrameWhenCryptorNotReady = keyProviderOptions[@"discardFrameWhenCryptorNotReady"];
+
+  NSNumber* keyDerivationAlgorithm = keyProviderOptions[@"keyDerivationAlgorithm"];
   
   RTCFrameCryptorKeyProvider* keyProvider =
       [[RTCFrameCryptorKeyProvider alloc] initWithRatchetSalt:ratchetSalt.data
@@ -361,7 +374,8 @@
                                          uncryptedMagicBytes: uncryptedMagicBytes != nil ? uncryptedMagicBytes.data : nil
                                             failureTolerance:failureTolerance != nil ? [failureTolerance intValue] : -1
                                                  keyRingSize:keyRingSize != nil ? [keyRingSize intValue] : 0
-                             discardFrameWhenCryptorNotReady:discardFrameWhenCryptorNotReady != nil ? [discardFrameWhenCryptorNotReady boolValue] : NO];
+                             discardFrameWhenCryptorNotReady:discardFrameWhenCryptorNotReady != nil ? [discardFrameWhenCryptorNotReady boolValue] : NO
+                                       keyDerivationAlgorithm:[self getKeyDerivationAlgorithm:keyDerivationAlgorithm]];
   self.keyProviders[keyProviderId] = keyProvider;
   result(@{@"keyProviderId" : keyProviderId});
 }