+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -6,9 +6,7 @@ import FlutterMacOS
 import Foundation
 
 import flutter_webrtc
-import path_provider_foundation
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   FlutterWebRTCPlugin.register(with: registry.registrar(forPlugin: "FlutterWebRTCPlugin"))
-  PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
 }